Mutant coverage in Status service object
Add mutant coverage to the Status service object.
