+++ app/controllers/tax_tribs/status_controller.rb
@@ -1,7 +1,9 @@
-class TaxTribs::StatusController < ApplicationController
-  respond_to :json
+module TaxTribs
+  class StatusController < ApplicationController
+    respond_to :json
 
-  def index
-    respond_with(Status.check)
+    def index
+      respond_with(Status.check)
+    end
   end
 end
